- 함대 아군, 적 나누는 시스템 변경 - 서버 점검중 기능 추가 - 튜토리얼 진행중

+++ Assets/Scripts/Space/Fleet/ModuleBase.cs
@@ -12,7 +12,7 @@ public class ModuleBase : MonoBehaviour
     [HideInInspector] public float m_healthMax;
     [HideInInspector] public float m_attack;
 
-    [HideInInspector] public long m_modulePointCostLevelup;
+    [HideInInspector] public int m_modulePointCostLevelup;
 
     // 리셋 시 환급할 투자 이력
     [HideInInspector] public int m_investedModulePoint;
